backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "FastAPI backend for ListingIQ: AI-powered insights for real estate listings with web scraping and comprehensive analysis."
authors = [
    { name = "Aarav Khanduja" }
]
requires-python = ">=3.13,<4.0"
license = { text = "MIT" }
dependencies = [
    "fastapi[standard]>=0.110.0,<1.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "ollama>=0.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "psutil>=5.9.0",
    "supabase>=2.0.0",
    "pydantic-settings>=2.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-mock>=3.10.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
uvicorn = { extras = ["standard"], version = "^0.29.0" }
black = "^25.1.0"
ruff = "^0.12.4"
pre-commit = "^4.2.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
